Use a worker to do the spell checking.
