+++ packages/cspell-eslint-plugin/cspell.json
@@ -5,6 +5,7 @@
     "words": [
         "estree",
         "pnpm",
+        "synckit",
         "treeshake",
         "tsbuildinfo"
     ]
