+++ packages/cspell-eslint-plugin/src/ASTNode.ts
@@ -1,0 +1,7 @@
+import type { Comment, Literal, Node } from 'estree';
+
+export interface JSXText extends Omit<Literal, 'type'> {
+    type: 'JSXText';
+}
+
+export type ASTNode = (Node | Comment | JSXText) & { parent?: Node };
